requirements.txt
django
coreapi
django-extensions
djangorestframework
django-rest-auth
django-allauth
sphinx
